+++ tests/Datasets/LabeledTest.php
@@ -637,13 +637,6 @@ class LabeledTest extends TestCase
         $this->assertEquals($expected, $stats);
     }
 
-    public function test_deduplicate() : void
-    {
-        $dataset = $this->dataset->deduplicate();
-
-        $this->assertCount(6, $dataset);
-    }
-
     public function test_transform_column() : void
     {
         $dataset = $this->dataset->transformColumn(3, 'abs');
@@ -667,5 +660,44 @@ class LabeledTest extends TestCase
         ];
 
         $this->assertEquals($expected, $desc);
+    }
+
+    public function test_deduplicate() : void
+    {
+        $dataset = $this->dataset->deduplicate();
+
+        $this->assertCount(6, $dataset);
+    }
+
+    public function test_to_array() : void
+    {
+        $expected = [
+            'samples' => self::SAMPLES,
+            'labels' => self::LABELS,
+        ];
+
+        $this->assertEquals($expected, $this->dataset->toArray());
+    }
+
+    public function test_to_json() : void
+    {
+        $expected = '{"samples":[["nice","furry","friendly",4],["mean","furry","loner",-1.5],'
+            . '["nice","rough","friendly",2.6],["mean","rough","friendly",-1],'
+            . '["nice","rough","friendly",2.9],["nice","furry","loner",-5]],'
+            . '"labels":["not monster","monster","not monster","monster","not monster","not monster"]}';
+            
+        $this->assertEquals($expected, $this->dataset->toJson());
+    }
+
+    public function test_to_csv() : void
+    {
+        $expected = 'nice,furry,friendly,4,not monster' . PHP_EOL
+            . 'mean,furry,loner,-1.5,monster' . PHP_EOL
+            . 'nice,rough,friendly,2.6,not monster' . PHP_EOL
+            . 'mean,rough,friendly,-1,monster' . PHP_EOL
+            . 'nice,rough,friendly,2.9,not monster' . PHP_EOL
+            . 'nice,furry,loner,-5,not monster' . PHP_EOL;
+            
+        $this->assertEquals($expected, $this->dataset->toCsv());
     }
 }